+++ src/js/controller.js
@@ -4,14 +4,6 @@ import recipeView from "./views/recipeView.js";
 import "core-js/stable";
 import "regenerator-runtime/runtime";
 
-const timeout = function (s) {
-  return new Promise(function (_, reject) {
-    setTimeout(function () {
-      reject(new Error(`Request took too long! Timeout after ${s} second`));
-    }, s * 1000);
-  });
-};
-
 // https://forkify-api.herokuapp.com/v2
 
 ///////////////////////////////////////
